requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bitarray==3.3.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
cytoolz==1.0.1
daphne==4.1.2
defusedxml==0.7.1
Django==5.0.4
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
gunicorn==22.0.0
h2==4.1.0
hexbytes==1.3.0
hpack==4.1.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.0
parsimonious==0.10.0
pillow==10.3.0
priority==1.3.0
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rlp==4.1.0
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.0
tomli==2.2.1
toolz==1.0.0
twilio==9.4.1
Twisted==24.11.0
txaio==23.1.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
web3==7.10.0
websockets==15.0.1
yarl==1.18.3
zope.interface==7.2
